+++ src/components/Form/Form.scss
@@ -1,15 +1,5 @@
 @import '../../styles/abstract/constants.scss';
 
-.form {
-  width: 692px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid red;
-  margin: 104px auto auto auto;
-}
-
 .form__container {
   display: flex;
   flex-direction: column;
@@ -31,26 +21,4 @@
       align-items: center;
     }
   }
-}
-
-.form__title {
-  align-self: flex-start;
-}
-
-.form__subtitle {
-align-self: flex-start;
-}
-
-.form__inputsContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  margin: 20px 0 40px 0;
-  width: 408px;
-}
-
-.form__checkboxContainer {
-  align-self: flex-start;
-  justify-self: flex-start;
-  margin-left: 3px;
 }
