+++ src/components/StylesForForms/Form.scss
@@ -1,0 +1,56 @@
+@import '../../styles/abstract/constants.scss';
+
+.form {
+  width: 692px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid red;
+  margin: 104px auto auto auto;
+}
+
+.form__container {
+  display: flex;
+  flex-direction: column;
+  margin: 100px auto;
+  justify-content: flex-start;
+  align-items: center;
+  &_kind {
+    &_register {
+      width: 408px;
+    }
+    &_login {
+      width: 408px;
+
+    }
+    &_recovery {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+    }
+  }
+}
+
+.form__title {
+  align-self: flex-start;
+}
+
+.form__subtitle {
+align-self: flex-start;
+}
+
+.form__inputsContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin: 20px 0 40px 0;
+  width: 408px;
+}
+
+.form__checkboxContainer {
+  align-self: flex-start;
+  justify-self: flex-start;
+  margin-left: 3px;
+}
